Move team splitting logic, implement basic team tile display
